Renamed action modes "daemon" and "unique" to "exclusive" and "queue", respectively
