+++ core/action-manager.js
@@ -1,7 +1,7 @@
 var modes = {
     "immediate": require('./immediate-action'),
-    "daemon": require('./daemon-action'),
-    "unique": require('./unique-action')
+    "exclusive": require('./exclusive-action'),
+    "queue": require('./queue-action')
 };
 
 function ActionManager(spec) {
