+++ core/exclusive-action.js
